requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.2.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.6.2
blis==0.7.9
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.4
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
cryptography==41.0.3
cycler==0.12.1
cymem==2.0.6
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.103.0
fastjsonschema==2.19.1
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.0
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
grpcio==1.48.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.10
keras==2.15.0
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.4.5
langcodes==3.3.0
libclang==18.1.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.0
mistune==3.0.2
ml-dtypes==0.2.0
multidict==6.0.4
murmurhash==1.0.7
mysql-connector-python==8.4.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.1.3
notebook_shim==0.2.4
numexpr==2.8.7
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.1
preshed==3.0.6
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.15.1
PyJWT==2.4.0
pyOpenSSL==23.2.0
pyparsing==3.1.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rpds-py==0.18.1
rsa==4.7.2
scikit-learn==1.4.2
scipy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_680rtkz_e8/croot/scipy_1701295052241/work/dist/scipy-1.11.4-cp39-cp39-macosx_11_0_arm64.whl#sha256=94e819573aa4910f68819f513bcdad27c0da5299f1bbf71b1ed7176288fdcde3
Send2Trash==1.8.3
shellingham==1.5.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.25
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
tensorboard==2.15.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.15.0
termcolor==2.1.0
terminado==0.18.1
thinc==8.2.2
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.29.0
wasabi==0.9.1
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.3.8
widgetsnbextension==4.0.10
wrapt==1.14.1
yarg==0.1.9
yarl==1.9.3
zipp==3.17.0
